Move zip_with_gaps to itertools
